Implement route index
Implement the route GET /index which returns the template
index.html.tera and consequently the other included templates in
frontend/templates/. The route is handled by routes::pages::index,
which builds the context and renders the template.
To enable the handler to build the context the locale middleware is
added, which adds a FluentBundle with the requests locale to
incoming requests. The struct TextMapBuilder is added in
routes::pages to simplify the construction of the context with the
necessary translations.
The function routes::create_path_map() is added to simplify the
creation of a PATHS attribute, which holds the paths of all assets
referenced by the templates and is therefore added to the context.